refactor: 쿠폰 선택 및 적용 로직 useCouponSelection 훅으로 분리
