+++ src/features/Cart/components/OrderConfirm.tsx
@@ -1,4 +1,6 @@
 import { css } from '@emotion/react';
+import { useCartInfo } from '../hooks/useCartInfo';
+import { useOrderInfo } from '../hooks/useOrderInfo';
 
 import { Button } from '@/shared/components/Button/Button';
 import { Flex } from '@/shared/components/Flex/Flex';
@@ -7,101 +9,32 @@ import { Text } from '@/shared/components/Text/Text';
 
 import Back from '../../../../public/Back.png';
 import { StepProps } from '@/shared/types/funnel';
-import { useOrderInfo } from '@/features/Cart/hooks/useOrderInfo';
 import { CartListContainer } from '../container/CartListContainer';
 import { CartItemDetail } from './CartItemDetail';
 import { RemoteAreaCheckBox } from './RemoteAreaCheckBox';
 import { PriceSummary } from './PriceSummary';
-import { useCartInfo } from '../hooks/useCartInfo';
-import { useContext, useEffect, useState } from 'react';
 import { CouponModal } from '@/features/Coupon/components/CouponModal';
-import { getCouponList } from '@/features/Coupon/api/coupon';
-import { useFetchData } from '@/shared/hooks/useFetchData';
-import { Coupon, CouponResponse } from '@/features/Coupon/types/Coupon.types';
-import { isError } from '@/shared/utils/isError';
-import { ToastContext } from '@/shared/context/ToastProvider';
-import { isCouponValid } from '@/features/Coupon/utils/validateCoupon';
-import { getBestCouponCombination } from '@/features/Coupon/utils/combinations';
-import { useCartContext } from '../context/CartProvider';
-import { usePriceInfo } from '../hooks/usePriceInfo';
-import { calculateTotalDiscount } from '@/features/Coupon/utils/calculateTotalDiscount';
+import { useCouponSelection } from '@/features/Coupon/hooks/useCouponSelection';
+
+import { Coupon } from '@/features/Coupon/types/Coupon.types';
 
 type OrderConfirmProps = {
   onSelectCoupons: (selected: Coupon[]) => void;
 } & StepProps;
 
 export const OrderConfirm = ({ onSelectCoupons, onPrev, onNext }: OrderConfirmProps) => {
-  const {cartItems } = useCartContext();
-  const { hasCheckCartLength, totalQuantity, totalPrice } = useOrderInfo();
   const { selectedCartItems } = useCartInfo();
-  const { showToast } = useContext(ToastContext);
-  const { isRemoteArea } = useCartContext();
-  const { deliveryFee } = usePriceInfo();
-  const [coupons, setCoupons] = useState<Coupon[]>([]);
-  const selectedCoupons = coupons.filter((c) => c.checked && !c.disabled);
-  const [showCouponList, setShowCouponList] = useState(false);
-  const [discountAmount, setDiscountAmount] = useState(0);
-  const coupon = useFetchData<CouponResponse[]>({ autoFetch: getCouponList });
+  const { hasCheckCartLength, totalQuantity } = useOrderInfo();
 
-  useEffect(() => {
-    if (coupon.data) {
-      const validated = coupon.data.map((c) => ({
-        ...c,
-        checked: false,
-        disabled: !isCouponValid(c, selectedCartItems, totalPrice),
-      }));
-      setCoupons(validated);
-    }
-  }, [coupon.data]);
-
-  useEffect(() => {
-    if (coupon.error && isError(coupon.error)) {
-      showToast('쿠폰 정보를 불러올 수 없습니다.');
-    }
-  }, [coupon.error, showToast]);
-
-  useEffect(() => {
-    if (showCouponList) {
-      const validCoupons = coupons.filter((c) => !c.disabled);
-      const best = getBestCouponCombination(validCoupons, cartItems, {
-        isRemoteArea,
-        totalPrice,
-        deliveryFee,
-      });
-      setCoupons((prev) =>
-        prev.map((c) => ({
-          ...c,
-          checked: best.some((b) => b.id === c.id),
-        }))
-      );
-    }
-  }, [showCouponList]);
-
-  const onToggleCoupon = (id: number) => {
-    setCoupons((prevCoupons) => {
-      const selectCouponCount = prevCoupons.filter((c) => c.checked).length;
-      const target = prevCoupons.find((c) => c.id === id);
-
-      if (!target) return prevCoupons;
-
-      if (!target.checked && selectCouponCount >= 2) return prevCoupons;
-
-      return prevCoupons.map((coupon) =>
-        coupon.id === id ? { ...coupon, checked: !coupon.checked } : coupon
-      );
-    });
-  };
-
-  const onApplyCoupons = () => {
-    const selected = coupons.filter((c) => c.checked && !c.disabled);
-    const discount = calculateTotalDiscount(selectedCartItems, selected, {
-      totalPrice,
-      isRemoteArea,
-    });
-
-    setDiscountAmount(discount);
-    setShowCouponList(false);
-  };
+  const {
+    coupons,
+    discountAmount,
+    showCouponList,
+    selectedCoupons,
+    setShowCouponList,
+    onToggleCoupon,
+    onApplyCoupons,
+  } = useCouponSelection();
 
   const handleApplyCoupons = () => {
     onSelectCoupons(selectedCoupons);
@@ -127,17 +60,17 @@ export const OrderConfirm = ({ onSelectCoupons, onPrev, onNext }: OrderConfirmPr
         width="100%"
         padding="40px 20px"
       >
-        <Text type="Heading" weight="bold">
-          주문 확인
-        </Text>
+        <Text type="Heading" weight="bold">주문 확인</Text>
         <Text type="Caption" weight="regular">
           {`총 ${hasCheckCartLength}종류의 상품 ${totalQuantity}개를 주문합니다.\n최종 결제 금액을 확인해 주세요.`}
         </Text>
+
         <CartListContainer variant="review">
           {selectedCartItems?.map((item) => (
             <CartItemDetail key={item.id} variant="review" {...item} />
           ))}
         </CartListContainer>
+
         <Button
           width="100%"
           size="xl"
@@ -154,9 +87,11 @@ export const OrderConfirm = ({ onSelectCoupons, onPrev, onNext }: OrderConfirmPr
         >
           쿠폰 적용
         </Button>
+
         <RemoteAreaCheckBox />
         <PriceSummary variant="review" discountAmount={discountAmount} />
       </Flex>
+
       <Button
         width="100%"
         size="xl"
@@ -168,6 +103,7 @@ export const OrderConfirm = ({ onSelectCoupons, onPrev, onNext }: OrderConfirmPr
       >
         결제하기
       </Button>
+
       {showCouponList && (
         <CouponModal
           coupons={coupons}
